* Singled out reusable part of spectroscopy test.
